requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
wandb>=0.15.0
datasketch>=1.5.9
pyahocorasick>=2.0.0
xxhash>=3.2.0
datasets>=2.14.0
accelerate>=0.21.0
tqdm>=4.65.0
scikit-learn>=1.3.0
